Refactor: ProfileSet 컴포넌트 이름 변경 및 파일 이동
